refactor(cli): extract assets command module
